Extract JSON builder logic into TrackingEventsAPI.
